Hackmidwest Elasticsearch  - Where To Live
This is the final commit for Where To Live which was the winner for best machine learning application at Hackmidwest.
It includes the app that connects to Elasticsearch api.

The Elasticsearch ML model uses the data.gov HPI file to find the highest low mean function anomolies to
predict the cheapest state to live in.

Contact biga.david@gmail.com for the model or more information.